Fix layout_map_mag_dec function
`GeoMag()` call expects to get coordinates in the order of: latitude,
longitude. This means that we should use first the `y()` method
of `new_point`, and then `x()`.

Currently that it's opposite, which causes the declination
calculated with the layout function to be different than
what the main part of the plugin calculates and what the
compass rose shows.

This commit fixes the call to `GeoMag()` method by passing the
coordinate in the proper order.